+++ crates/base/alloc/src/data/codec/mod.rs
@@ -1,0 +1,17 @@
+// devela_base_alloc::data::codec
+//
+#![doc = crate::_DOC_DATA_CODEC!()] // public
+#![doc = crate::_doc!(modules: crate::data; codec: hash)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: hash)]
+//
+
+// mod bit; // WIP
+
+crate::structural_mods! { // _mod
+    _mods {
+        // pub use super::{
+        //     bit::_all::*,
+        // };
+    }
+}