+++ crates/base/alloc/src/data/layout/linked/_reexport.rs
@@ -1,8 +1,0 @@
-// devela_base_alloc::data::layout::linked::_reexport
-
-use crate::{_TAG_DATA_STRUCTURE, _reexport};
-
-_reexport! { rust: alloc::collections, location: "data/layout", tag: _TAG_DATA_STRUCTURE!(),
-    doc: "A doubly-linked list with owned nodes.",
-    LinkedList
-}