+++ crates/base/alloc/src/data/layout/mod.rs
@@ -3,10 +3,10 @@
 #![doc = crate::_DOC_DATA_LAYOUT!()] // public
 #![doc = crate::_doc!(modules: crate::data; layout: array)] // space
 #![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: array, collections, vec)]
 #![doc = crate::_QUO_DATA_LAYOUT!()]
 //
 
-mod linked;
 // mod pool;
 // mod queue;
 mod sort; // Sort
@@ -18,7 +18,6 @@ pub mod array;
 crate::structural_mods! { // _mods, _pub_mods
     _mods {
         pub use super::{
-            linked::_all::*,
             // pool::_all::*,
             // queue::_all::*,
             sort::_all::*,