+++ crates/base/alloc/src/data/mod.rs
@@ -9,17 +9,19 @@
 #![cfg_attr(feature = "safe_data", forbid(unsafe_code))]
 // docs
 crate::CONST! { pub(crate) _DOC_DATA_MODULES =
-    crate::_doc!(modules: crate; data: key, layout, uid); // access, codec, error, iter
+    crate::_doc!(modules: crate; data: id, layout, topol); // access, codec, error, value
 }
 
 pub mod id;
 pub mod layout;
+pub mod topol;
 
 crate::structural_mods! { // _pub_mods, _crate_internals
     _pub_mods {
         pub use super::{
             id::_all::*,
             layout::_all::*,
+            topol::_all::*,
         };
     }
     _crate_internals {