+++ crates/base/alloc/src/data/topol/linked/_reexport.rs
@@ -1,0 +1,8 @@
+// devela_base_alloc::data::topol::linked::_reexport
+
+use crate::{_TAG_DATA_STRUCTURE, _reexport};
+
+_reexport! { rust: alloc::collections, location: "data/topol", tag: _TAG_DATA_STRUCTURE!(),
+    doc: "A doubly-linked list with owned nodes.",
+    LinkedList
+}