+++ crates/base/alloc/src/data/topol/linked/mod.rs
@@ -1,0 +1,15 @@
+// devela_base_alloc::data::topol::linked
+//
+#![doc = crate::_DOC_DATA_TOPOL_LINKED!()] // private
+#![doc = crate::_doc!(modules: crate::data::topol; linked)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: collections)]
+//
+
+mod _reexport; // SYMLINK from /src/data/topol/linked/_reexport_alloc.rs
+
+crate::structural_mods! { // _mods, _reexports
+    _reexports {
+        pub use super::_reexport::*;
+    }
+}