+++ crates/base/core/src/data/layout/mod.rs
@@ -3,11 +3,11 @@
 #![doc = crate::_DOC_DATA_LAYOUT!()] // public
 #![doc = crate::_doc!(modules: crate::data; layout: array)] // dst
 #![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: array, collections, vec)]
 #![doc = crate::_QUO_DATA_LAYOUT!()]
 //
 
 mod buffer; // buffer_linear!
-mod linked;
 // mod pool;
 // mod queue;
 mod sort; // Sort
@@ -20,7 +20,6 @@ crate::structural_mods! { // _mods, _pub_mods
     _mods {
         pub use super::{
             buffer::_all::*,
-            linked::_all::*,
             // pool::_all::*,
             // queue::_all::*,
             sort::_all::*,