+++ crates/base/core/src/data/topol/linked/mod.rs
@@ -1,7 +1,7 @@
-// devela_base_core::data::layout::linked
+// devela_base_core::data::topol::linked
 //
-#![doc = crate::_DOC_DATA_LAYOUT_LINKED!()] // private
-#![doc = crate::_doc!(modules: crate::data::layout; linked)]
+#![doc = crate::_DOC_DATA_TOPOL_LINKED!()] // private
+#![doc = crate::_doc!(modules: crate::data::topol; linked)]
 #![doc = crate::_doc!(flat:"data")]
 #![doc = crate::_doc!(extends: collections)]
 //