+++ crates/base/core/src/data/topol/mod.rs
@@ -4,9 +4,11 @@
 #![doc = crate::_DOC_DATA_TOPOL!()] // public
 #![doc = crate::_doc!(modules: crate::data; topol)] // graph, spatial
 #![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: collections)]
 #![doc = crate::_QUO_DATA_TOPOL!()]
 //
 
+mod linked;
 // mod ord;
 // mod span;
 
@@ -15,10 +17,11 @@
 
 crate::structural_mods! { // _mods, _pub_mods
     _mods {
-        // pub use super::{
-        //     // ord::_all::*,
-        //     // span::_all::*,
-        // };
+        pub use super::{
+            linked::_all::*,
+            // ord::_all::*,
+            // span::_all::*,
+        };
     }
     _pub_mods {
         // pub use super::{